requirements.txt
azure-core==1.26.3
azure-cosmos==4.3.1
fastapi==0.92.0
pydantic==1.10.5
python-dotenv==1.0.0
requests==2.28.2
uvicorn==0.20.0
